requirements.txt
aiohttp==3.7.4.post0
async-timeout==3.0.1
attrs==21.4.0
certifi==2021.10.8
cffi==1.15.0
chardet==4.0.0
charset-normalizer==2.0.9
discord==1.7.3
discord.py==1.7.3
idna==3.3
multidict==5.2.0
prettytable==2.5.0
pycparser==2.21
PyNaCl==1.4.0
python-dotenv==0.19.2
requests==2.26.0
six==1.16.0
typing_extensions==4.0.1
urllib3==1.26.7
wcwidth==0.2.5
yarl==1.7.2
youtube-dl==2021.12.17
